move file results navigator into list service
