+++ src/vs/workbench/parts/debug/electron-browser/callStackView.ts
@@ -24,9 +24,8 @@ import { IWorkspaceContextService, WorkbenchState } from 'vs/platform/workspace/
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 import { Source } from 'vs/workbench/parts/debug/common/debugSource';
 import { basenameOrAuthority } from 'vs/base/common/resources';
-import { WorkbenchTree } from 'vs/platform/list/browser/listService';
+import ResourceResultsNavigation, { WorkbenchTree } from 'vs/platform/list/browser/listService';
 import { IWorkbenchEditorService } from 'vs/workbench/services/editor/common/editorService';
-import FileResultsNavigation from 'vs/workbench/parts/files/browser/fileResultsNavigation';
 
 const $ = dom.$;
 
@@ -105,9 +104,9 @@ export class CallStackView extends TreeViewsViewletPanel {
 				twistiePixels
 			});
 
-		const fileResultsNavigation = new FileResultsNavigation(this.tree);
+		const fileResultsNavigation = new ResourceResultsNavigation(this.tree);
 		this.disposables.push(fileResultsNavigation);
-		this.disposables.push(fileResultsNavigation.openFile(e => {
+		this.disposables.push(fileResultsNavigation.openResource(e => {
 			if (this.ignoreSelectionChangedEvent) {
 				return;
 			}
